Renamed `crate::instrument_memory` to `crate::instrument`.
